+++ src/QUAD7SHIFT.h
@@ -56,7 +56,7 @@
                 uint8_t reverseBits(uint8_t);
                 void transferDigit(uint16_t);
                 unsigned int fixedPointRepresentation(float,uint8_t);
-                void printNumber(uint8_t, uint8_t);
+                void printNumber(uint16_t, uint8_t);
                 uint8_t getTableIndex(char);
                 void printDigit( uint8_t , uint8_t , bool  );
